requirements.txt
faiss-cpu 
sentence-transformers 
fastapi 
uvicorn 
streamlit 
python-dotenv 
sqlalchemy 
requests
dotenv
openai